+++ activetigger/models.py
@@ -28,7 +28,7 @@ class BertModel():
 
     """
 
-    def __init__(self, path:str) -> None:
+    def __init__(self, path:Path) -> None:
         """
         All the data are sorted in path/bert/$NAME
         """
